backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
fastapi==0.115.12
frozenlist==1.6.0
gitdb==4.0.12
gitpython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm-watson-machine-learning==1.0.367
ibm-watsonx-ai==1.3.13
idna==3.10
importlib-metadata==8.7.0
jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-ibm==0.3.10
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8
langgraph==0.4.0
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.64
langgraph-supervisor==0.0.20
langsmith==0.3.38
lomond==0.3.3
markupsafe==3.0.2
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.1.0
narwhals==1.37.1
numpy==1.26.4
ollama==0.4.8
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pandas==2.1.4
pillow==11.2.1
pip==25.1
propcache==0.3.1
protobuf==6.30.2
pyarrow==20.0.0
pydantic==2.11.4
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.2
requests-toolbelt==1.0.0
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlalchemy==2.0.40
starlette==0.46.2
streamlit==1.45.0
streamlit-autorefresh==1.0.1
tabulate==0.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing-extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
